+++ api/query/table.go
@@ -7,6 +7,7 @@ package query
 
 import (
 	"fmt"
+	"sort"
 	"strings"
 	"time"
 )
@@ -146,9 +147,10 @@ func NewFluxRecord(table int, values map[string]interface{}) *FluxRecord {
 	return &FluxRecord{table: table, values: values}
 }
 
-// Table returns index of the table record belongs to
+// Table returns value of the table column
+// It returns zero if the table column is not found
 func (r *FluxRecord) Table() int {
-	return r.table
+	return int(intValue(r.values, "table"))
 }
 
 // Start returns the inclusive lower time bound of all records in the current table.
@@ -204,14 +206,22 @@ func (r *FluxRecord) ValueByKey(key string) interface{} {
 
 // String returns FluxRecord string dump
 func (r *FluxRecord) String() string {
-	var buffer strings.Builder
+	if len(r.values) == 0 {
+		return ""
+	}
+
 	i := 0
-	for k, v := range r.values {
-		if i > 0 {
-			buffer.WriteString(",")
-		}
-		buffer.WriteString(fmt.Sprintf("%s:%v", k, v))
+	keys := make([]string, len(r.values))
+	for k := range r.values {
+		keys[i] = k
 		i++
+	}
+	sort.Strings(keys)
+	var buffer strings.Builder
+	buffer.WriteString(fmt.Sprintf("%s:%v", keys[0], r.values[keys[0]]))
+	for _, k := range keys[1:] {
+		buffer.WriteString(",")
+		buffer.WriteString(fmt.Sprintf("%s:%v", k, r.values[k]))
 	}
 	return buffer.String()
 }
@@ -227,7 +237,7 @@ func timeValue(values map[string]interface{}, key string) time.Time {
 	return time.Time{}
 }
 
-// timeValue returns string value from values map according to the key
+// stringValue returns string value from values map according to the key
 // Empty string is returned if key is not found
 func stringValue(values map[string]interface{}, key string) string {
 	if val, ok := values[key]; ok {
@@ -236,4 +246,15 @@ func stringValue(values map[string]interface{}, key string) string {
 		}
 	}
 	return ""
+}
+
+// intValue returns int64 value from values map according to the key
+// Zero value is returned if key is not found
+func intValue(values map[string]interface{}, key string) int64 {
+	if val, ok := values[key]; ok {
+		if i, ok := val.(int64); ok {
+			return i
+		}
+	}
+	return 0
 }